requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
deap==1.4.1
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.42.1
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.58.0
h5py==3.9.0
humanfriendly==10.0
idna==3.4
imageio==2.31.3
imageio-ffmpeg==0.4.9
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
keras==2.11.0
keyboard==0.13.5
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.0
Pillow==10.0.1
protobuf==3.19.6
pyaml==23.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.24.0
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
SQLAlchemy==2.0.20
sympy==1.12
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tifffile==2023.8.30
torch==2.0.1
torch-tools==0.1.5
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
visdom==0.2.4
websocket-client==1.6.3
Werkzeug==2.3.7
wrapt==1.15.0
